+++ backend/canisters/group_index/impl/src/model/public_groups.rs
@@ -101,7 +101,6 @@ impl PublicGroups {
             frozen: None,
             events_ttl: summary.events_ttl,
             events_ttl_last_updated: summary.events_ttl_last_updated,
-            gate: summary.gate_config.as_ref().map(|gc| gc.gate.clone()),
             gate_config: summary.gate_config,
             wasm_version: BuildVersion::default(),
         })