+++ backend/canisters/group_index/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Removed
+
+- Remove `gate` which is superseded by `gate_config` ([#8186](https://github.com/open-chat-labs/open-chat/pull/8186))
+
 ## [[2.0.1790](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1790-group_index)] - 2025-06-13
 
 ### Changed